# Conflicts:
#	code/game/objects/items/stacks/tiles/tile_types.dm
#	code/game/turfs/unsimulated/planetary_vr.dm
#	vorestation.dme

+++ code/modules/vore/eating/digest_act_vr.dm
@@ -9,6 +9,11 @@
 			var/obj/item/device/pda/P = src
 			if(P.id)
 				P.id = null
+
+		for(var/mob/living/voice/V in possessed_voice) // Delete voices.
+			V.Destroy() //Destroy the voice.
+		for(var/mob/living/M in contents)//Drop mobs from objects(shoes) before deletion
+			M.forceMove(item_storage)
 		for(var/obj/item/O in contents)
 			if(istype(O, /obj/item/weapon/storage/internal)) //Dump contents from dummy pockets.
 				for(var/obj/item/SO in O)
@@ -17,6 +22,7 @@
 					qdel(O)
 			else if(item_storage)
 				O.forceMove(item_storage)
+		GLOB.items_digested_roundstat++
 		qdel(src)
 		return w_class
 
@@ -37,6 +43,10 @@
 			var/obj/item/device/pda/P = src
 			if(P.id)
 				P.id = null
+		for(var/mob/living/voice/V in possessed_voice) // Delete voices.
+			V.Destroy() //Destroy the voice.
+		for(var/mob/living/M in contents)//Drop mobs from objects(shoes) before deletion
+			M.forceMove(item_storage)
 		for(var/obj/item/O in contents)
 			if(istype(O,/obj/item/weapon/storage/internal)) //Dump contents from dummy pockets.
 				for(var/obj/item/SO in O)
@@ -45,7 +55,16 @@
 					qdel(O)
 			else if(item_storage)
 				O.forceMove(item_storage)
-		qdel(src)
+		if(istype(src,/obj/item/stack))
+			var/obj/item/stack/S = src
+			if(S.get_amount() <= 1)
+				qdel(src)
+			else
+				S.use(1)
+				digest_stage = w_class
+		else
+			GLOB.items_digested_roundstat++
+			qdel(src)
 	if(g_damage > w_class)
 		return w_class
 	return g_damage
@@ -63,7 +82,7 @@
 	return FALSE
 /obj/item/weapon/pinpointer/digest_act(var/atom/movable/item_storage = null)
 	return FALSE
-/obj/item/blueprints/digest_act(var/atom/movable/item_storage = null)
+/obj/item/areaeditor/blueprints/digest_act(var/atom/movable/item_storage = null)
 	return FALSE
 /obj/item/weapon/disk/nuclear/digest_act(var/atom/movable/item_storage = null)
 	return FALSE
@@ -82,7 +101,8 @@
 		icon = 'icons/obj/card_vr.dmi'
 		icon_state = "[initial(icon_state)]_digested"
 	else
-		sprite_stack += "digested"
+		if(!sprite_stack.Find("digested"))
+			sprite_stack += "digested"
 	update_icon()
 	return FALSE
 
@@ -110,8 +130,7 @@
 /obj/item/organ/digest_act(atom/movable/item_storage = null)
 	if((. = ..()))
 		if(isbelly(item_storage))
-			var/obj/belly/B = item_storage
-			. += 2 * (B.digest_brute + B.digest_burn + (B.digest_oxy)/2)
+			. *= 3
 		else
 			. += 30 //Organs give a little more
 
@@ -126,6 +145,10 @@
 /////////////
 /obj/item/device/mmi/digital/posibrain/digest_act(atom/movable/item_storage = null)
 	//Replace this with a VORE setting so all types of posibrains can/can't be digested on a whim
+	return FALSE
+
+/obj/item/organ/internal/nano/digest_act(atom/movable/item_storage = null)
+	//Make proteans recoverable too
 	return FALSE
 
 // Gradual damage measurement
